backend/app/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
python-multipart==0.0.20
mediapipe==0.10.21
numpy==1.26.4
pillow==11.1.0